requirements.txt
asgiref==3.8.1
boto3==1.36.3
botocore==1.36.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-storages==1.14.4
djangorestframework==3.15.2
idna==3.10
jmespath==1.0.1
mysqlclient==2.2.7
pillow==11.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
s3transfer==0.11.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
